feat(friends): made gateways talk with service then socket broadcast
